Update compare_subcompartment_size_distribution.py
Plots produced by the previous impl. were a bit misleading, as they
placed a lot of emphasis on small (but somewhat frequent)
subcompartments.
This commit changes the logic so that plots show the number of bps
covered by a subcompartment that spans X bps.

Example:
With the older impl. a subcompartment with extent 10kbp had the same
weight as a subcompartment of extent 100kbp.
In the current implementation, subcompartments of extent 100kbp have 10x
the weight of subcompartments with extent of 10kbp.